requirements.txt
Django==4.1.3
djangorestframework==3.14.0
Markdown==3.4.1
django-filter==22.1
django-crispy-forms==1.14.0
python-dotenv==0.21.0
psycopg2-binary==2.9.5
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.13.0